1) Added means to install services to start at boot up. 2) A clumsy attempt to make main.py more robust for when the network is slow on boot up.

+++ Code/setup.py
@@ -1,7 +1,8 @@
 import os
 import sys
 import time
-os.system("cd /usr/bin && sudo rm python && sudo ln -s python3 python")
+if sys.version.startswith('2'):
+    os.system("cd /usr/bin && sudo rm python && sudo ln -s python3 python")
 flag=0x00
 for x in range(1,4):
     if os.system("sudo apt-get update") == 0:
@@ -25,3 +26,13 @@ if flag==0x0F:
 else:
     print ("\nSome libraries have not been installed yet. Please run 'sudo python setup.py' again")
 
+
+
+if True: # Work-In-Progress: APH added this section to make the server start with boot up
+    os.system("sudo ln -s ./Scripts/start-robot.sh /usr/sbin/start-robot.sh")
+    os.system("sudo ln -s ./Scripts/robot.service /etc/systemd/system/robot.service")
+    os.system("sudo systemctl daemon-reload")
+    os.system("sudo systemctl enable robot")
+    os.system("sudo systemctl start robot")
+    if os.system("sudo systemctl status robot") != 0:
+        print("\n Failed to configure robot server to start on boot up.")
